Fix visual issues with RelativePositionSelect
Now the selector actually snaps to things properly! I changed the snap lines to not extend to the edges of the control anymore, which made some calculations a bit easier. I suspect that the visual issues occurred due to my calculations being off somehow.

Also added a few more options to the test panel in SsuiSample.