Change some properties to dependency properties
Didn't realize that SnapToSnapLines, SnapDistance, and KeyMoveStep weren't already dependency properties. Now they are!

The remaining properties in RelativePositionSelect (being SelectorSize, SelectedWidth, and SelectedHeight) aren't going to be as easy to convert to dependency properties. I can see myself making it happen with SelectorSize, but not the others at this stage.