Create TwoPositionsSelect control, for use later on
It's like the RelativePositionSelect control... but with two selectors! Oooooh!

I figured this would be a more useful control to use for the linear gradient editor, so that both the Start Point and End Point could be visible in one control.

There's only two key differences between this control and the original RelativePositionSelect: 1) a selector can only be moved by mouse by clicking and dragging it, rather than being able to click anywhere in the control. 2) with keyboard support, you use the Shift key to switch which selector is being moved. (In the future, I may change which key is being used, other than Shift).

Otherwise, practically all of RelativePositionSelect's functionality was copied!

Finally, did some minor code organization with RelativePositionSelect, and added in a proper SelectedPoint property.