First start with LinearGradientEditorDialog
Put together the basic controls and such needed, and hooked them up just enough to show that the basics are working. Right now, the dialog more just "shows" the gradient, not really allowing it to be edited, but this is just a start. I'll also add labels and rearrange some stuff too as time goes on.