Add VersionEditor, exception handling in CreateEditorForType
Added a VersionEditor property editor, for System.Version objects. May not be the most useful and needed, but I wanted to do it anyway lol. Includes the ability to parse a string.

Added exception handling around CreateEditorForType, in case issues occur while trying to load in the editor for a property type. In all cases, "null" is returned.