Redesigned Add method handing in ListEditor
Redesigned how the Add function works in ListEditor. Now when opening the ListEditor, the ListEditor will check against various things to try to allow new items to be added to the list.

Now, if the base object has a parameterless constructor or a constructor that takes 1 string as a parameter, it should work. There is also special handling for most built-in C# types.