Add Multiline String Editor dialog, for use in StringEditor
Now if editing a string in the Property List, you can use a Multiline Editor to more easily put in new lines and tabs. Use Ctrl+Enter to accept the input and close the dialog.

Originally, I had put the dialog under the PropertyList Dialogs namespace, but I moved it to the Utils namespaces for slightly more exposure. I might actually move it to a new Dialogs namespace in SSUI 2.0.

Also added the ability to view a list as an IEnumerable of char, which is... less useful than I thought. So I also disabled the menu item for that lol.

This did give me the chance to correct some things with the Enumerable warning in the List Editor Dialog. :)