Fix bug about updating items in ListEditorDialog
When trying to edit an item in a list, the ListEditorDialog was still holding on to the original item's reference. This meant that if the item was updated/replaced (such as via an earlier edit), then it's possible the reference the dialog still has isn't good anymore, and the dialog can't find the new item anymore to update it again.

Thus, I had to change the core eventargs class to store the old value of any updated property, so that I can give the editor access to the most recent version of that item.

With items that are mutable, this was probably never an issue before. For immutable items, I'm not sure how I didn't come across this earlier.