Add in CharEditor, CharInputDialog for PropertyList
This was one of the property types I was excited about adding in, and doing proper justice. As I went along, I learned about the existence of the Rune type (only in .NET/.NET Core), so this editor supports both chars and Runes. I've had to use a lot of #if statements to separate the Rune stuff for .NET only.

The editor in the property list control will display the char/Rune when possible, and also displays the hex Unicode number representing that char/Rune. You can type in any character to set a new char, although surrogates will lock the text box and display "(surrogate)" in it.

You also have the option to select a char/Rune from the new CharInputDialog, where you type or select the Unicode code point for the char/Rune you want. If you locked the text box with a surrogate char, this will allow you to put in a non-surrogate char afterwards.

I can imagine this editor maybe being a bit annoying for those trying to enter in chars not in the basic multilingual Unicode plane, but for here and now, I feel this is the best I can easily do. If there's other ideas about how to visit this, I'll definitely entertain them.