BrushEditor - limit options based upon property type
Only display all brush options when the property type is "Brush"; otherwise, limit options based upon the specific types of brushes that are allowed by the property. This prevents the BrushEditor from supplying, for example, a SolidColorBrush, in places that ask for an ImageBrush.

Not sure how many situations this will be applicable in, but I think it's better to be robust!