Fix stack overflow while trying to round values
When in a situation where the MaxValue has more decimals supplied than the Decimals property allows, and the rounded value is larger than the MaxValue, the DoubleSpinner gets stuck in a loop trying to apply the MaxValue and then rounding it. So now I check and limit the value before any further work is done.