Add Angle functionality in Linear Gradient Dialog
Now changing the Angle value does something! Also the Angle value will be updated when you change the points independently.

Some of the angle calculating code (angle to points) comes from Microsoft, but I had to modify it so that the end results matched the limits of the TwoPositionsSelect control. Other angle calculating code (points to angle) comes from StackOverflow, yay!

Also updated IsPropertyWritable in the BrushEditor so that all available editing buttons/controls are disabled if the property is read-only.