Fix images being too small on buttons lol
The Padding value in FlatButton is meant to reflect the Padding value for the standard WPF button, but in situations like this, it makes unintended results lol.

Don't know if I want to try to make any corrections about this at the FlatButton control level or not, about dynamically decreasing the padding as the button gets smaller, as I feel that's going to be a lot of work for something relatively minor.