Add DisplayAsHex option to IntegerSpinner
I was worried how much time and code it would take to make this happen, but surprisingly... not very much!

When set to true, DisplayAsHex will display its value as a hexadecimal number (i.e. "12" turns to "C"), and also AcceptExpressions is disabled. This can be useful when interacting with bytes or computer representations of numbers or data.

With this, I added a "Display as Hex" option to the IntegerEditor for the PropertyList control.