Add SwatchesResetTransparency property to dialog
By default, this is set to true. This was added to aid with the user experience; if the transparency of the "current" color was 0, then clicking on a swatch seemingly made no change as the transparency was still 0. Now, this will reset the transparency to 255 (all other tabs will not reset the transparency, although maybe I'll add options for those later).

There may be situations where one would want the transparency to not change, and that is why I made this an option, rather than enforcing a default.