Added SelectedTabContent dependency property
Added a SelectedTabContent dependency property to TabControl, which is used internally to set the main content of the control now (rather than directly in C# like before). Also changed SelectedTab to a dependency property.

I was hoping this would fix the issue we were seeing with the ColorPickerDialog's buttons not working with binding, but that doesn't seem to be the case here.