Clicking on window icon will now display system menu
When you left click or right click on the window icon, the system menu (with options like Move, Size, Maximize, and Close) will appear below it, as is usual with standard window behavior.

I actually implemented this as a flexible function, in case for some weird reason consumers of this library want to display the menu in some other weird spot.