Remove "unsafe" code, obsolete StreamExtensions
Marked the entire StreamExtensions file as obsolete, as I don't actually use it anywhere in SSUI and mainly kept it around in case others used it. I fully removed the function ReadSingleBigEndian, as it was the only function in the entire package that used the "unsafe" code block; by removing it, I could disable the use unsafe csproj setting.