Cleanup SsuiSample project file
Continuing on earlier work to reduce the amount of cruft in the project files. Ideally, now it should be much nicer looking!

Also removed .NET Core 3.1 target for SsuiSample, as I'll be dropping support for 3.1 and 5 in a later version of SSUI.

Finally, minor code improvement in FlatWindowTest, as suggested by the analyzer.