Refactor ChromeButtons code, add CaptionButtonPadding
Ever since I took on the original FlatWindow code and expanded upon it, I never really dabbled much in the ChromeButtons code as it never caused problems and I had no reason to mess with it. But now I've taken the time to actually look through, figure out what all the code is doing, and remove a lot of weirdly redundant or useless bits.

I fixed some visual glitches along the way, and also exposed a new property called ButtonPadding (or CaptionButtonPadding in the FlatWindow). This can be used to make the buttons bigger or smaller as needed.