Add Windows11Interop project... which fails build
Adds a Windows11Interop project, which is meant to be a companion to SSUI to quickly enable a developer to use Windows 11 caption buttons for enabling Snap Layouts... but for whatever reason, the project instantly fails to build due to a "runtime pack" error. I've filed some bug reports about this online, so hopefully I get some resolution.

Anyway... uhhh, this is a thing, I guess.