Added note about ImageColorPicker with DPI awareness
ImageColorPicker works in programs that have DPI awareness set; otherwise, if on a screen with a DPI other than 100%, it will look and act weirdly.

I had made a mention about this in the release notes (way back in 1.7), but there wasn't an explicit note about it here in the code.