Properly handle RTL systems with ShowSystemMenu
On RTL systems, the system menu was appearing in the top-left corner rather than the top-right, so now this rectifies that.

It's interesting to note that SSUI looks decently well in RTL systems with no extra work on my end! (I think it's because WPF does all the heavy lifting.) I was worried how much custom code I'd have to make to properly support it, but it turns out not much at all!