Display 16x16 icon in FlatWindow
Used and expanded upon a converter I found on StackOverflow; when you provide it a size value, it'll attempt to load an icon of that size (or the nearest size available).

Now the top-left icon displayed in FlatWindow will be the 16x16 one, rather than the 32x32 one or a different one like before.