Change SelectPanel behavior by item selection method
Add function called SetIsSelectedWithSource which is used to determine how the item was selected. This is used to determine what actions the SelectPanel should take.

Now clicking on the checkboxes can be used to select multiple items in SelectPanel, even if the Ctrl key isn't held down.

In the future, I might be able to use this as a basis for other things too, we'll see.