Add DisplayOptions to PropertyList, add attributes
The DisplayOptions property and the new PropertyListHide and PropertyListShow attributes can be used to control whether properties should be hidden in the PropertyList. Hidden properties aren't even loaded at all.

The PropertyListDisplayFlags enum is a flags enum, so multiple flags can be set to hide multiple types of properties.

I also added a UI option to hide read-only properties.