Reorganize list description text in ListEditorDialog
Reorganized the description text for the list or IEnumerable opened in the dialog, to sound/look a little less clunky and wordy.

Unfortunately, this is another one of those things that is hardcoded English and I don't have a good way right now to allow changing that into another language.