ReloadObject persists Name, fix DisplayOptions enum
Made the ReloadObject overload with the resetViewSettings option publicly-accessible. Also, the ObjectDisplayName is also preserved; previously when clicking on the Reload button, the custom set name would be lost if one was provided.

Also changed the enum value of ShowAll, as the value of 0 was screwing up the display calculations.