Add LongSpinner and LongEditor
The LongSpinner is practically identical to the IntegerSpinner, except it stores and displays numbers as a long (Int64) rather than an int (Int32).

Added a LongEditor to PropertyList that uses the LongSpinner; if a ulong value is entered that is larger than long.MaxValue, the value is displayed as text with the option for the user to change it to a lower value to allow editing.