Add DecimalEditor, fix DoubleEditor rounding
Added an editor that is used... for decimals. Makes sense given its name. The decimal editor primarily uses a DoubleSpinner for quick edits, but there's a "Set Exact Value" menu option for situations that go beyond what a double supports.

As part of this, I realized that both DecimalEditor and DoubleEditor didn't support numbers with more than 2 decimal digits, as the default DoubleSpinner decimal value is 2; thus, I changed these both to the max value of 15 so that more doubles and decimals are supported.