Remove commented out code from DoubleSpinner
Keeping all this old code around is doing no good for anyone, especially since I'm highly unlikely to come back to using whatever that code was.

In a future commit, I'll go through and remove commented out sections from IntegerSpinner and LongSpinner too.

Despite the fact that DoubleSpinner is technically based off a control that someone else made, a lot of it was rebuilt and redone by me, and the commented out sections are remnants from me experimenting with how to get certain things to work.