Add ShowGridlines and GridlineBrush to PropertyList
To allow the PropertyList to look a bit more like the WinForms PropertyGrid control, I added support for displaying gridlines. For now, I've also added menu items under the View menu of the PropertyList, but those may not stay there.

In future commits, I'll work on adding more brush properties for other aspects of the control.