Gridlines visual fixes, rewrite Types column display
Fixed the gridlines not appearing with the Types column, and fixed the "Show Gridlines" menu item not being checked when it should've been. Despite my original plans, I think my current goal is to leave these menu items in there (or maybe offer an option to hide them).

Rewrote the code for displaying the types of properties, to now instead display type names without displaying the full namespaces, and to more properly support generic types. The full names, including namespaces, are visible if you mouse over the type names.