Hook up brush properties in PropertyList
Now developers can modify the appearance of the PropertyList control, without having to use the ColorScheme property. There's a lot of brushes to control various aspects of the control, so users can make it exactly what they want.

I also overwrote the Content property of PropertyList to avoid developers accidentally messing with it.

Also minor rearrangement of code in FlatButton and StringEditor.