Add events, label properties to PropertyList
Added PropertyValueChanged and LoadedObjectChanged events to the PropertyList control, so developers can hook into those.

Also added a bunch of various label properties that developers can use to change the labels of various text strings in the control (for, say, localization).

I also changed PrettifyPropertyType into a public static method, so that I can utilize it anywhere and others might be able to take advantage of it too.