Add LoadFromType function to CommandKeyAction
This function will create CommandKeyActions from the static properties of a passed-in type. This is another convenience function, similar to the LoadFromMenu function in RoutedEventKeyAction. The idea is to be able to load all commands from, say, ApplicationCommands or NavigationCommands, which then can be easily used to connect keyboard shortcuts to.

The actual function was included in the previous commit by accident.