Update version to 1.9.6 - commit no. 500!
This is the 500th commit to GitHub! Wow! And along with this commit, I've updated the package version to 1.9.6. It's time for release! :)

The next 500 commits and beyond will be the start of a new journey for Solid Shine UI. I'll be excited to see where this goes. Here's to another 500! :)