Beginning work on Ribbon control
Implemented the basic back-end logic for most of the classes that will make up the Ribbon, and started a little bit with designing the templates. Still a decent amount of work to go, but this is a solid start on this journey, I feel.

In the future, I'll need to spend a bit more time deciding how I'm going to implement the resizing/collapsing aspect of the Ribbon; the properties present right now probably don't mesh together well.