Set up RibbonTest pane in SsuiSample
Set up the RibbonTest pane to display a RibbonGroup and... allow testing it. Once I have the full Ribbon control up and running, I'll use this pane for displaying that instead. For now, this confirms that RibbonGroup is working, as well as RibbonButton and RibbonMenuButton.

Also added some icons to use with the buttons!