Add RibbonTwoLineText from WPF Ribbon repo
Rather than building my own solution to do the same thing, I instead took the RibbonTwoLineText control from WPF's Ribbon and implemented it in my own code. I did tweak it a bit to 1) enable nullability support and 2) allow it to interact with MenuButton's ShowMenuArrow property.

Adding this control has the added benefit of being able to display the menu arrow on the bottom row of text similar to Microsoft's own Ribbon implementations, rather than the (comparatively janky-looking) method I had of simply putting the arrow always on the right.

For those who always want the arrow on the right for some reason, I enabled that behind MenuButton's KeepMenuArrowOnRight property. In the future, I'll also look to implement that property when RibbonMenuButton is in Small or IconOnly modes.