Set a MaxHeight for RibbonGroup (to 96)
This way, even when RibbonGroup is used outside of a Ribbon, it still maintains the expected shape. Allowing a smaller height allows situations like displaying a Ribbon without group headers and/or only a single row of commands... but there's more that will need to be done to fully enable that all the way through.

Developers can still change the MaxHeight (or set it to NaN) as needed.