Initial progress with Ribbon, RibbonTab visuals
Made some initial progress with visuals for the Ribbon and RibbonTab. RibbonTab will use a similar setup to the TabItem from my TabControl, so a lot of the code added here was copied over and modified from TabItem, TabControl, and TabDisplayItem.

There's still a decent amount of work to do, but this is one more step!