Finish Ribbon visual design, add tab selection
Finished the design of the Ribbon control itself, and now users are able to actually fill in tabs and groups and they appear on the Ribbon! There's still a looooot of progress to go yet (especially around selecting/switching tabs), a lot of things aren't quite working right yet, but progress is being made!