Re-organize/re-layout LinearGradientEditorDialog
Reorganized the code for the LinearGradientEditorDialog. No functionality was changed, just changed the order of some functions in the code and put it all under some #region groups.

Also changed the margins/spacing of various components, to fix the weirdly big gap between the gradient bar and the rest of the dialog, and the relatively small space the OK/Cancel buttons occupied in the corner.