Create non-generic ISelectableCollectionSource
SelectPanel's selection handling wasn't working properly, if the ItemsSource wasn't set to exactly the expected item type... so I created a non-generic interface for ISelectableCollectionSource, which ends up taking care of this problem. I just kind of wish this problem didn't occur in the first place, but I suspect this is a deeper C# problem (or something that I don't have enough understanding/research yet to solve in a different way).

In the future, I'm thinking of setting this interface to take from IEnumerable, and then from there, it'll be easier for other developers to utilize this particular interface and go from there.