Further work on TransformEditDialog
A lot of the visual parts are now working, with being able to add on new transforms, and displaying the appropriate editing pane when one is selected... but right now, nothing gets loaded or saved. So the dialog is pretty, but ultimately useless at the moment.

Still more work to go!