Start with Transform handling in BrushEditor
First commit for adding support for the Transform and RelativeTransform properties of brushes. Clicking on the buttons open up the TransformEditDialog, but right now, the dialog doesn't actually load in anything and nothing occurs when you hit Save.

Once the dialog is finished though, you'll be able to use it for editing these properties. I'll also use the dialog in other places too! :D