SelectPanel MoveItemUp/Down works with ItemsSource
Now if you set the ItemsSource property of SelectPanel, the MoveItemUp and MoveItemDown functions will still work (as long as ItemsSource is still an IList). SelectPanel will also maintain selection if ItemsSource is an ISelectableCollectionSource.

The convenience functions MoveSelectedItemsDown and MoveSelectedItemsUp still only work when the developer is working with Items.