Further work on TransformEditDialog
Made the window a bit smaller (as the existing height was way too much). Added Move Up and Move Down buttons for the list on the left. Also added in a MatrixTransform editor, and rewrote how the editors are loaded. Finally, I also changed all of the DoubleSpinners to have a Step of 1.