Rename ISelectableCollectionSource to ISelectableCollection
I changed the interfaces to take from IEnumerable and ICollection<T> respectively, and with that, renamed them to ISelectableCollection since they are now guaranteed to be implemented on top of a collection.

Made some modifications to SelectableCollectionView to support ICollection<T>. The Clear method is not supported right now, as I'd instead recommend developers just interact with the base collection to perform that method.

Also changed CollectionSelectionChangedEventArgs to use ICollection rather than IEnumerable, so that a Count property is guaranteed to be exposed.