Create derived ItemsControl to use in RibbonGroup
Created a modified version of the ItemsControl that allows me to modify how items are rendered on the RibbonGroup, beyond what was doable before. With this, I was able to make it so narrower buttons didn't take up the full width of a column when not needed.

Ideally with only this one small little change, everything else should continue to act as normal. But over time, we'll see if there's any knock-on effects I need to be wary of.