Store transform edits in TransformEditDialog
Arguably the most important part of having a dialog for editing an item, now the edits will actually stick! This can also be extracted via the new ExportTransforms method. With this, the dialog is at least now useable!

A secondary mode, called Single Edit Mode, will need to be added to the dialog before it can be considered 100% complete, but this is a major piece of progress!