Add File menu control (still a work in progress!)
Added in a File menu control, which takes from ButtonBase but doesn't take from FlatButton (so I had to recreate some of the mouse handling code that was in FlatButton). Clicking on the File menu button will automatically open the menu itself, but right now the actual contents of the menu isn't fully built out yet.

In the future, I'm thinking of creating IFileMenuRibbonControl versions of certain controls, which won't be fun, and ideally the better solution would be creating unique styles for IRibbonControl elements added onto the file menu. So I'll have to think about how I want to do that.

But at least it's very good and solid progress! :D One more big thing done for the Ribbon control!