Create templated control version of PropertyList
This is going to be the same, full PropertyList control, but now as a templated control, rather than a UserControl!

Right now, most of the functionality is there and all of the visuals are there, but I'm not 100% sure if everything is working properly. I haven't tested anything yet, so that'll be my next task.

But for now, I sleep!