Further work on new PropertyList control
I think most missing functionality is now present... but now the question is trying to get the functionality that's present to actually... function.

To that end, the SsuiSample app now uses this new PropertyList control, which I'll be using extensively in the future in order to fix bugs and improve the control.