Fixing up things in new PropertyList control
Fixed the binding issues and other issues I was having with the new PropertyList control, and now the new control is almost a complete drop-in for the old control! Still some bugs and other stuff to fix, though. But at this point, it's almost all the way there!

I might consider replacing the standard WPF menu in the View menu with a more custom drop-down popup instead...