Various Improvements to BrushEditor
A lot of various improvements to the BrushEditor.

First of all, Edit Transform and Edit Relative Transform are now both links, rather than headers with buttons. Takes up less space.

Also rewrote a lot of how brushes are stored and presented outside of the control. Now, when GetValue is called, a copy of the brush is made and provided to the caller, rather than the direct internal brush value itself.

Hooked up the transform edit dialog to now apply its values to the brushes... but it doesn't seem to work? I'll need to investigate further.