Finish work on TransformEditor
Refactored how GetValue and LoadValue works, by only checking for the type as needed. Also I make sure that I make clones of the transform values, to avoid situations where I end up with frozen Transforms that I cannot edit.

I also made sure that the TransformEditDialog is only in Single Edit Mode if only a specific type is needed.

Finally, I added some special rendering if it's an identity matrix, since that seems to be the default value in WPF.

There may be some bugs that I may still need to iron out, but I think this is actually pretty solid. Good to go!