Add TransformEditor, improve GetEditorForType
Added in the TransformEditor as an editor for transforms in PropertyList. I also reworked the GetEditorForType function by allowing and adding in fallback values if an editor for certain basic types aren't registered.

As a part of this, I added a new extension method! The "DependencyObjectExtensions" file is now renamed to "BaseExtensions". This is done by mainly just copying the GetValueOrDefault present in newer .NET versions, but with a minor tweak.