Improve layout of BrushEditor's transform section
I decided to remove the shadows and move the Edit Transform links to the right side. I changed the text color to match the foreground color to make sure it matched all color schemes, but I do worry that it might've made it a bit harder to recognize as being click-able. I might revisit this.

Either way, the edit panel now looks a decent bit cleaner!