Various improvements to various PropertyEditors
Fix TransformEditor and DoubleEditor still allowing editing of read-only properties.

When the CornerRadiusEditor and ThicknessEditor are below a certain width, the arrows on the DoubleSpinners will be hidden (so that the numbers are still visible). Also added "+1" and "-1" options in the ThicknessEditor's menu.

Minor refactoring of the LongEditor.

Finally, for all the editors I touched on today, I also moved setting the ColorScheme to a method, rather than using the property. My eventual goal is to remove the set-only ColorScheme property and instead use the method in the IPropertyEditor interface.