Introduce generic IconVariation enum
This replaces the MessageDialogImageColor enum, which was in a weird spot and I also had other uses for such an enum in other places in my library.

In the future, I'll be doing more in regards to handling icons and icon variations. But this is a great step in a direction I want.