Improvements to Ribbon visuals/contextual tab handling
Added on proper brush properties for the Ribbon, so that we can use these rather than binding to the ColorScheme. With these, I also fixed the visuals so that the colors are what I intend them to be.

In the future, when I rework the ColorScheme, I'll need to go through and update this as well. But that'll just be a part of that process.

I also increased the default padding for each TabItem to be 12, to better match what is used in Office and other Microsoft Ribbon controls.

Finally, I made sure that contextual tab support is working as expected. For the most part it is, although for some reason I couldn't get binding to work. I added on functionality to automatically select another tab if the contextual tab is hidden (similar to what is done in Office).