Fix up Ribbon tab drag-drop... and then disabling it
I figured out why tab drag-drop wasn't working... it was because I had commented out the code that would've made it work. Now that code is working.

However, I've now set AllowControlReordering to false by default, so that dragging and dropping doesn't happen. This isn't an area I'm expecting to be able to build a lot in during 2.0, so it'll remain turned off by default, until I visit it more in later versions.