Remove commented out code in FlatButton
There is a loooot of old commented-out code in FlatButton, that I'm just not going to need anymore. A lot of it was from prior to the change to using a control template, which isn't needed now.

I also removed the obsolete method ApplyColorScheme(HighContrastOption) from FlatButton, MenuButton, and ContextMenu. I'll continue to remove it from more controls in future commits.