Rename "AsThis" method to "PerformAs"
I figure that "PerformAs" will make a bit more sense to newcomers to the code. I also added a new overload that adds on an extra parameter, which will do a type-check on that parameter.

With that new overload, I changed the TransparentBack and UseAccentColors properties in FlatButton to use this new method. This technically results in less lines of code, although it might be a bit harder to grasp as a newcomer.