Implement SplitButton, other minor fixes
Implemented the SplitButton control! This combines a FlatButton and MenuButton together, so that you can click the larger FlatButton to perform the main action, or the smaller MenuButton to open a menu with more options. It works surprisingly well, and has most of the features of a FlatButton (thanks to good ol' copy and paste).

A few notes:
1. There's actually not a MenuButton for the menu part. It's another FlatButton and SplitButton copied all of MenuButton's code.
2. There's still some bugs to fix and visual improvements to make.
3. One of the improvements already made though is through the PartialCornerRadiusConverter. It's a lengthy name, but helps make the CornerRadius property work well on a SplitButton.

Some other changes in here include adding some more comments to FlatButton.