Add separator bar to SplitButton, finish template
Rather than having a weird thick border between the main and menu buttons of the SplitButton, now I removed that border and instead replaced it with a narrower and more controllable separator bar.

To make these separator bar visuals work, I added on a PartialThicknessConverter. I also modified the PartialCornerRadiusConverter to support a byte parameter, to help save memory.

The separator bar can be changed in height and color, but width is currently not working? Bug to fix later!

Also not working is the new MenuButtonPlacement property. This should change where around the main button that the menu button is placed, but changing this value seems to do nothing. I'll have to see if it's something to do with default values being done in the template directly, rather than the triggers.

Other than that, the SplitButton is pretty much complete!