Fix SeparatorThickness in SplitButton
The SeparatorThickness property now works! However, to be honest, it starts to look kind of bad if you set it to any value above 3, so it's probably best to just leave it as is. But why not offer the option, I guess lol.

If this ends up bugging me too much (and feeling too pointless), I may remove it later. We'll see.