Add ColorSwatchButton, more X11 color helper methods
Created a ColorSwatchButton control for the purpose of making an easy to add on button for selecting a color. I tried this out on a few buttons in the ColorPickerDialog and I'm liking the results, so this will be rolled out to the rest of the dialog's Swatch tab in a future commit.

This will also prove useful for a future control I have in mind.

I also added some more helper methods to ColorsHelper, for directly getting names from the X11 color table and listing out the entire color table.