Replace all swatch buttons with ColorSwatchButtons
Replaced all the swatch buttons in the ColorPickerDialog with ColorSwatchButtons. The functionality should be practically identical, but it is a bit easier code for me to manage.

Also, to more properly support transparency, I've updated how the ColorSwatchButton actually displays its color (so now, a checkerboard background is used with semi-transparent colors).