Clean up SelectableUserControl, add IClickSelectableControl
Cleaned up and improved on the code of SelectableUserControl, including adding some dependency properties and routed events. Added a new SelectOnClick property, and improved XML comments.

Also added the IClickSelectableControl interface, which I think will be useful for some kind of later application (not sure what yet lol). This includes a loooot of the base properties and methods that the relevant controls use. I'll implement it in more controls and add more interfaces on as needed.