Beginning work on Ribbon Gallery
The first steps for building the Gallery control. Nothing visual is happening yet, but the underlying properties and data is put into place, so that I can work on it more later.

This includes the massive GalleryItemLayout enum lol. I tried to put in examples for each layout type of where I've found them in Microsoft Office, which hopefully helps other developers visualize what that layout type will look like.