Further refactoring of spinner controls
I've managed to move a lot more functionality into the SpinnerBase class, by heavy usage of virtual methods.

I don't think I'm going to move ahead with creating a generic version of SpinnerBase, as I think I'll have a hard time getting that to rectify with WPF's dependency property system in a way that I'll be satisfied with.

By using virtual methods to allow the bulk of the code to be in SpinnerBase, and having the individual classes just for UI interop and type-specific functionality, I think I'll end up with some clean and nice code.