Add BeginButtonPress event to SpinnerBase
Yet another method to help simplify and consolidate all of the spinner code. With this, now most button related functions are performed via activating a SpinnerBase function.

The biggest thing still being done in the individual spinner controls is UI-related changes and any code specific to that spinner's data type.