Add CornerRadiusModifierConverter
This will be used in the future. This converter adds two values together to create a new CornerRadius.

I wonder if there is a different/better way to do this in WPF, but as far as I'm aware, there isn't, so this is what I'm doing lol

Anyway, you'll see this used in the near future