Begin next phase for SpinnerBase: NewSpinnerBase
My next goal for the spinner controls is to make control templates for these. I originally opted to use UserControls for the spinner controls as it was just easier for me to directly access and affect the subelements, but now I think using control templates will be a better option for extensibility -similar to when I changed FileSelect to use a control template.

As a part of that, I've had to create a copy of SpinnerBase and IntegerSpinner. The eventual goal is that these will fully replace the existing controls, but that'll happen in a future commit.