Further work on Gallery control - added visuals!
Implemented the control templates for the Gallery and GalleryItem, and so now both render on the Ribbon! There's still a lot of bugs and issues to sort through, including the arrow icons on the Gallery's buttons, but it's one major step in the right direction!

I've set up GalleryItem to implement from IClickSelectableControl, which I've found helpful to ensure that I set up my controls in a way that implement those features.

For the future, I'll need to mess with the "Filter" function that I've created in Gallery, which I'll use to actually set up the items that are displayed in the in-Ribbon Gallery, and from there, add support for scrolling.

The next next step after that will be creating the Gallery drop-down, and that'll be its own whole thing to figure out. I'll need to do some thinking about that.