Implement basic functionality in NewIntegerSpinner
Added functionality for the textbox in NewIntegerSpinner. Still need to do more to get the buttons to actually function, and I still need to finish setting up the visuals, but at least it's somewhere.

Also added in MaxValueChanged and MinValueChanged events to NewSpinnerBase, and rewrote how value validation is done. Ideally, we should only need to check that MaxValue and MinValue are valid only when those are changed.

From here, the next steps are to integrate support for the commands, and then use connect them up to the buttons... somehow lol. Given that the buttons operate on checking when the clicking begins and then stops, rather than a singular Click event, I may have to give some thought to how I will implement this.